backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
ccxt==4.1.92
pandas==2.1.1
numpy==1.24.3
ta==0.10.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
requests==2.31.0
